+++ src/xrGame/Hit.h
@@ -59,18 +59,18 @@ struct SHit
 	void Write_Packet(NET_Packet& P);
 	void Write_Packet_Cont(NET_Packet& P);
 
-	void GenHeader(u16 PacketType, u16 ID);
+	void GenHeader(u16 PacketType, u32 ID);
 	//private:
 	//GE_HIT
 	u32 Time;
 	u16 PACKET_TYPE;
-	u16 DestID;
+	u32 DestID;
 
 	float power;
 	Fvector dir;
 	CObject* who;
-	u16 whoID;
-	u16 weaponID;
+	u32 whoID;
+	u32 weaponID;
 	u16 boneID;
 	Fvector p_in_bone_space;
 	float impulse;
