+++ src/xrGame/relation_registry.h
@@ -93,8 +93,8 @@ public:
 	struct FIGHT_DATA
 	{
 		FIGHT_DATA();
-		u16 attacker;
-		u16 defender;
+		u32 attacker;
+		u32 defender;
 		float total_hit;
 		u32 time;
 		u32 time_old;
@@ -116,7 +116,7 @@ public:
 	};
 
 	//зарегистрировать драку (реакция на Hit в EntityAlive)
-	void FightRegister(u16 attacker, u16 defender, ALife::ERelationType defender_to_attacker, float hit_amount);
+	void FightRegister(u32 attacker, u32 defender, ALife::ERelationType defender_to_attacker, float hit_amount);
 	void UpdateFightRegister();
 
 private:
@@ -124,7 +124,7 @@ private:
 	static FIGHT_VECTOR* m_fight_registry;
 	static FIGHT_VECTOR& fight_registry();
 
-	FIGHT_DATA* FindFight(u16 object_id, bool by_attacker/* = true*/);
+	FIGHT_DATA* FindFight(u32 object_id, bool by_attacker/* = true*/);
 	static RELATION_MAP_SPOTS* m_spot_names;
 public:
 	const shared_str& GetSpotName(ALife::ERelationType& type);
