+++ src/xrServerEntities/xrServer_Object_Base.h
@@ -72,9 +72,9 @@ public:
 	u16 m_script_version;
 	u16 RespawnTime;
 
-	u16 ID; // internal ID
-	u16 ID_Parent; // internal ParentID, 0xffff means no parent
-	u16 ID_Phantom; // internal PhantomID, 0xffff means no phantom
+	u32 ID; // internal ID
+	u32 ID_Parent; // internal ParentID, 0xffffffff means no parent
+	u32 ID_Phantom; // internal PhantomID, 0xffffffff means no phantom
 	xrClientData* owner;
 
 	// spawn data
@@ -83,7 +83,7 @@ public:
 	GameTypeChooser m_gameType;
 	u8 s_RP;
 	Flags16 s_flags; // state flags
-	xr_vector<u16> children;
+	xr_vector<u32> children;
 
 	// update data
 	Fvector o_Position;
